refactor: move voxie specific player code to voxie module
