+++ src/systems/mod.rs
@@ -2,8 +2,6 @@
 pub mod gun;
 pub mod physics;
 #[cfg(feature = "gui")]
-pub mod player;
-#[cfg(feature = "gui")]
 pub mod projectiles;
 #[cfg(feature = "gui")]
 pub mod skybox;
