+++ src/voxie/mod.rs
@@ -1,2 +1,3 @@
 pub mod game_context;
+pub mod player;
 pub mod scene;
